refactor(orga): remove FT userPixOrgaNewAuthDesign
