+++ orga/app/controllers/join/signup.js
@@ -1,16 +1,9 @@
 import Controller from '@ember/controller';
-import { service } from '@ember/service';
 
 export default class SignupController extends Controller {
-  @service featureToggles;
-
   queryParams = ['code', 'invitationId'];
   code = null;
   invitationId = null;
-
-  get isNewAuthDesignEnabled() {
-    return this.featureToggles.featureToggles.usePixOrgaNewAuthDesign;
-  }
 
   get routeQueryParams() {
     return { code: this.code, invitationId: this.invitationId };
