+++ orga/app/styles/app.scss
@@ -45,10 +45,8 @@
 @use 'components/previous-page-button' as *;
 @use 'components/pix-score-tag' as *;
 @use 'components/progress-bar' as *;
-@use 'components/register-form' as *;
 @use 'components/team' as *;
 @use 'components/team/invite-form' as *;
-@use 'components/login-or-register' as *;
 @use 'components/manage-authentication-method-modal' as *;
 @use 'components/participation-filters' as *;
 @use 'components/learner/header' as *;
