--- orga/app/styles/components/register-form.scss
@@ -1,45 +1,0 @@
-@use 'pix-design-tokens/typography';
-@use 'pix-design-tokens/fonts';
-
-.register-form {
-  max-width: 360px;
-
-  &__cgu-label {
-    margin: 0;
-    font-size: 0.875rem;
-    font-family: fonts.$font-roboto;
-  }
-
-  &__cgu-error {
-    margin-top: 20px;
-    color: var(--pix-error-700);
-    font-size: 0.75em;
-  }
-
-  &__information {
-    @extend %pix-body-s;
-
-    color: var(--pix-neutral-900);
-    text-align: center;
-  }
-
-  .input-container {
-    margin: var(--pix-spacing-4x) 2px var(--pix-spacing-4x) 2px;
-  }
-
-  .pix-input-password > .pix-input__error-message {
-    bottom: calc(-18px - 1px - 0.75rem);
-  }
-
-  .checkbox-container {
-    display: flex;
-
-    input[type='checkbox'] {
-      margin: 1px 6px 0 0;
-    }
-  }
-
-  .cgu-validation-message {
-    margin-top: 20px;
-  }
-}
