+++ orga/app/templates/authentication/login.gjs
@@ -1,60 +1,29 @@
-import PixBlock from '@1024pix/pix-ui/components/pix-block';
 import t from 'ember-intl/helpers/t';
 import pageTitle from 'ember-page-title/helpers/page-title';
 import AuthenticationIdentityProviders from 'pix-orga/components/authentication/authentication-identity-providers';
 import LoginForm from 'pix-orga/components/authentication/login-form';
 import AuthenticationLayout from 'pix-orga/components/authentication-layout/index';
-import LocaleSwitcher from 'pix-orga/components/locale-switcher';
-import PageTitle from 'pix-orga/components/ui/page-title';
 
 <template>
   {{pageTitle (t "pages.login.title")}}
-  {{#if @controller.isNewAuthDesignEnabled}}
-    <AuthenticationLayout class="signin-page-layout">
-      <:header>
-      </:header>
 
-      <:content>
-        <div>
-          <h1 class="pix-title-m">{{t "pages.login.title"}}</h1>
-          <h3 class="pix-body-l">{{t "pages.login.with-pix-account"}}</h3>
-        </div>
+  <AuthenticationLayout class="signin-page-layout">
+    <:header>
+    </:header>
 
-        <LoginForm
-          @isWithInvitation={{false}}
-          @hasInvitationAlreadyBeenAccepted={{@controller.hasInvitationAlreadyBeenAccepted}}
-          @isInvitationCancelled={{@controller.isInvitationCancelled}}
-        />
-
-        <AuthenticationIdentityProviders />
-      </:content>
-    </AuthenticationLayout>
-
-  {{else}}
-    <main class="login-page">
+    <:content>
       <div>
-        <PixBlock class="login-page__container" @variant="orga">
-          <header class="login-page__header">
-            <img src="/pix-orga-color.svg" alt="Pix Orga" />
-            <PageTitle @centerTitle={{true}}>
-              <:title>
-                {{t "pages.login.title"}}
-              </:title>
-            </PageTitle>
-
-          </header>
-
-          <LoginForm
-            @isWithInvitation={{false}}
-            @hasInvitationAlreadyBeenAccepted={{@controller.hasInvitationAlreadyBeenAccepted}}
-            @isInvitationCancelled={{@controller.isInvitationCancelled}}
-          />
-        </PixBlock>
-
-        {{#if @controller.isInternationalDomain}}
-          <LocaleSwitcher />
-        {{/if}}
+        <h1 class="pix-title-m">{{t "pages.login.title"}}</h1>
+        <h3 class="pix-body-l">{{t "pages.login.with-pix-account"}}</h3>
       </div>
-    </main>
-  {{/if}}
+
+      <LoginForm
+        @isWithInvitation={{false}}
+        @hasInvitationAlreadyBeenAccepted={{@controller.hasInvitationAlreadyBeenAccepted}}
+        @isInvitationCancelled={{@controller.isInvitationCancelled}}
+      />
+
+      <AuthenticationIdentityProviders />
+    </:content>
+  </AuthenticationLayout>
 </template>
