--- orga/tests/integration/components/auth/login-or-register-test.js
@@ -1,120 +1,0 @@
-import { clickByName, render } from '@1024pix/ember-testing-library';
-import { click } from '@ember/test-helpers';
-import { hbs } from 'ember-cli-htmlbars';
-import { t } from 'ember-intl/test-support';
-import { module, test } from 'qunit';
-import sinon from 'sinon';
-
-import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
-
-module('Integration | Component | Auth::LoginOrRegister', function (hooks) {
-  setupIntlRenderingTest(hooks);
-
-  let loginButton;
-
-  hooks.beforeEach(function () {
-    const featureToggles = this.owner.lookup('service:featureToggles');
-    sinon.stub(featureToggles, 'featureToggles').value({ usePixOrgaNewAuthDesign: false });
-    loginButton = t('pages.login-or-register.login-form.button');
-  });
-
-  test('displays the organization name the user is invited to', async function (assert) {
-    // when
-    const invitationMessage = t('pages.login-or-register.title', { organizationName: 'Organization Aztec' });
-
-    await render(hbs`<Auth::LoginOrRegister @organizationName='Organization Aztec' />`);
-
-    // then
-    assert.dom('.login-or-register-panel__invitation').hasText(`${invitationMessage}`);
-  });
-
-  test('toggles the register form by default', async function (assert) {
-    // when
-    await render(hbs`<Auth::LoginOrRegister />`);
-
-    // then
-    assert.dom('.register-form').exists();
-  });
-
-  test('toggles the login form on click on login button', async function (assert) {
-    // given
-    await render(hbs`<Auth::LoginOrRegister />`);
-
-    // when
-    await clickByName(loginButton);
-
-    // then
-    assert.dom('.login-form-legacy-design').exists();
-  });
-
-  test('toggles the register form on click on register button', async function (assert) {
-    // given
-    const registerButtonLabel = t('pages.login-or-register.register-form.button');
-
-    await render(hbs`<Auth::LoginOrRegister />`);
-
-    // when
-    await clickByName(loginButton);
-    await clickByName(registerButtonLabel);
-
-    // then
-    assert.dom('.register-form').exists();
-  });
-
-  module('when domain is not .fr', function () {
-    test('displays the locale switcher and translate to selected locale', async function (assert) {
-      // given
-      const domainService = this.owner.lookup('service:currentDomain');
-      sinon.stub(domainService, 'getExtension').returns('org');
-
-      const routerService = this.owner.lookup('service:router');
-      sinon.stub(routerService, 'replaceWith').returns(false);
-
-      // when
-      const screen = await render(hbs`<Auth::LoginOrRegister @organizationName='Organization Aztec' />`);
-
-      // then
-      await click(screen.getByRole('button', { name: t('components.locale-switcher.label') }));
-      await screen.findByRole('listbox');
-      await click(screen.getByRole('option', { name: 'English' }));
-      assert.dom(screen.getByText('You have been invited to join the organisation Organization Aztec')).exists();
-    });
-
-    test('saves selected locale and removes lang from query params', async function (assert) {
-      // given
-      const domainService = this.owner.lookup('service:currentDomain');
-      sinon.stub(domainService, 'getExtension').returns('org');
-
-      const routerService = this.owner.lookup('service:router');
-      const localeService = this.owner.lookup('service:locale');
-
-      sinon.stub(routerService, 'replaceWith');
-      sinon.stub(localeService, 'setCurrentLocale');
-
-      // when
-      const screen = await render(hbs`<Auth::LoginOrRegister @organizationName='Organization Aztec' />`);
-      await click(screen.getByRole('button', { name: t('components.locale-switcher.label') }));
-      await screen.findByRole('listbox');
-      await click(screen.getByRole('option', { name: 'English' }));
-
-      // then
-      assert.ok(localeService.setCurrentLocale.calledWithMatch('en'));
-      assert.ok(routerService.replaceWith.calledWithMatch({ queryParams: { lang: null } }));
-    });
-  });
-
-  module('when domain is .fr', function () {
-    test('does not display the locale switcher', async function (assert) {
-      // given
-      const domainService = this.owner.lookup('service:currentDomain');
-      sinon.stub(domainService, 'getExtension').returns('fr');
-
-      // when
-      const screen = await render(hbs`<Auth::LoginOrRegister @organizationName='Organization Aztec' />`);
-
-      // then
-      assert.dom(screen.getByText("Vous êtes invité(e) à rejoindre l'organisation Organization Aztec")).exists();
-      assert.dom(screen.queryByRole('button', { name: t('components.locale-switcher.label') })).doesNotExist();
-    });
-  });
-});
